LPS-174758 LPS-200418 portal-search: Move Base classes to new package and remove duplicated code
